Code review

+++ app/src/main/java/com/example/chatapp/data/repository/MessagesRepository.kt
@@ -15,30 +15,37 @@ class MessagesRepository(
     private val api: IAuthApi,
     private val source: IMessagesDataStore,
 ) : IMessagesRepository {
-    override suspend fun getAllMessages(): Result<Flow<List<Message>>> = withContext(Dispatchers.IO) {
-        tryLoadMessagesFromDataStore().let { result ->
-            source.fetchAllMessages().map { messages ->
-                messages.map { messageEntity ->
-                    messageEntity.toMessage()
-                }
-            }.let { flowResult ->
-                when (result) {
-                    is Result.Success -> {
-                        Result.Success(flowResult, result.message)
+    override suspend fun getAllMessages(): Result<Flow<List<Message>>> =
+        withContext(Dispatchers.IO) {
+            tryLoadMessagesFromDataBase().let { result ->
+                source.fetchAllMessages().map { messages ->
+                    messages.map { messageEntity ->
+                        messageEntity.toMessage()
                     }
-                    is Result.Error -> {
-                        Result.Error(flowResult, result.message)
+                }.let { flowResult ->
+                    when (result) {
+                        is Result.Success -> {
+                            Result.Success(flowResult, result.message)
+                        }
+                        is Result.Error -> {
+                            Result.Error(flowResult, result.message)
+                        }
                     }
                 }
             }
         }
-    }
-    override suspend fun tryLoadMessagesFromDataStore(): Result<Unit> {
+
+    override suspend fun tryLoadMessagesFromDataBase(): Result<Unit> {
         return api.getAllMessages().let { result ->
             when (result) {
                 is Result.Success -> {
+                    source.getCountMessages().let { tableSize ->
+                        if (tableSize != result.data?.size) {
+                            updateAllMessages(result.data ?: emptyList())
+                        }
+                    }
                     result.data?.map { message ->
-                        insertMessage(message = message)
+                        insertMessageFromDataBase(message = message)
                     }
                     Result.Success(message = "Подключено")
                 }
@@ -48,21 +55,29 @@ class MessagesRepository(
             }
         }
     }
-    override suspend fun insertMessage(message: Message) {
-        source.insertMessage(message = message.toEntity())
-    }
-
-    override suspend fun updateMessage(message: Message) {
-        source.updateMessage(message = message.toEntity())
-    }
-
     override suspend fun deleteAllMessages(): Result<DeleteAllMessagesResponse> =
         withContext(Dispatchers.IO) {
             try {
-                source.deleteAllMessages()
-                api.deleteAllMessages()
+                api.deleteAllMessages().let { response ->
+                    if (response.data!!.successful) {
+                        source.deleteAllMessages()
+                    }
+                    response
+                }
             } catch (e: Exception) {
                 Result.Error(message = e.message)
             }
         }
+    override suspend fun updateAllMessages(messages: List<Message>) {
+        source.deleteAllMessages()
+        messages.map { message ->
+            insertMessageFromDataBase(message = message)
+        }
+    }
+    override suspend fun insertMessageFromDataBase(message: Message) {
+        source.insertMessage(message = message.toEntity())
+    }
+    override suspend fun updateMessageFromDataBase(message: Message) {
+        source.updateMessage(message = message.toEntity())
+    }
 }
